feat: factor out parseFrameMessage
